use float field

+++ seqr/models.py
@@ -131,8 +131,8 @@ class ModelWithGUID(models.Model):
 
 class UserPolicy(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, primary_key=True)
-    privacy_version = models.DecimalField(max_digits=4, decimal_places=2, null=True)
-    tos_version = models.DecimalField(max_digits=4, decimal_places=2, null=True)
+    privacy_version = models.FloatField(null=True)
+    tos_version = models.FloatField(null=True)
 
 
 class Project(ModelWithGUID):
